% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auxdat_mecont}
\alias{auxdat_mecont}
\title{A simulated auxiliary data to show how to use 'bndovbme' function with continuous proxy variables}
\format{
A data frame with 3000 rows and 5 variables:
\describe{
\item{w1}{A common covariate in both main and auxiliary data}
\item{x}{A common covariate in both main and auxiliary data}
\item{z1}{A continuous proxy variable}
\item{z2}{A continuous proxy variable}
\item{z3}{A continuous proxy variable}
}
}
\source{
This dataset was simulated by simulatePackageData.R in data-raw folder
}
\usage{
auxdat_mecont
}
\description{
A simulated auxiliary data to show how to use 'bndovbme' function with continuous proxy variables
}
\keyword{datasets}
