% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auxdat_medisc}
\alias{auxdat_medisc}
\title{A simulated auxiliary data to show how to use 'bndovbme' function with discrete proxy variables}
\format{
A data frame with 3000 rows and 5 variables:
\describe{
\item{w1}{A common covariate in both main and auxiliary data}
\item{x}{A common covariate in both main and auxiliary data}
\item{z1}{A discrete proxy variable}
\item{z2}{A discrete proxy variable}
\item{z3}{A discrete proxy variable}
}
}
\source{
This dataset was simulated by simulatePackageData.R in data-raw folder
}
\usage{
auxdat_medisc
}
\description{
A simulated auxiliary data to show how to use 'bndovbme' function with discrete proxy variables
}
\keyword{datasets}
