% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auxdat_nome}
\alias{auxdat_nome}
\title{A simulated auxiliary data to show how to use 'bndovb' function}
\format{
A data frame with 50000 rows and 3 variables:
\describe{
\item{x1}{An omitted variable in the main data}
\item{x2}{A common covariate in both main and auxiliary data}
\item{x3}{A common covariate in both main and auxiliary data}
}
}
\source{
This dataset was simulated by simulatePackageData.R in data-raw folder
}
\usage{
auxdat_nome
}
\description{
A simulated auxiliary data to show how to use 'bndovb' function
}
\keyword{datasets}
