% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bndovb.R
\name{bndovb}
\alias{bndovb}
\title{bndovb}
\usage{
bndovb(maindat, auxdat, depvar, ovar, comvar, method = 1)
}
\arguments{
\item{maindat}{Main data set}

\item{auxdat}{Auxiliary data set}

\item{depvar}{A name of a dependent variable in main dataset}

\item{ovar}{A name of an omitted variable in main dataset which exists in auxiliary data}

\item{comvar}{A vector of common regressors existing in both main data and auxiliary data}

\item{method}{CDF and Quantile function estimation method.
Users can choose either 1 or 2. If the method is 1, the CDF and quantile function is estimated assuming a parametric normal distribution.
If the method is 2, the CDF and quantile function is estimated using a nonparaemtric estimator in Li and Racine(2008), Li, Lin, and Racine(2013).
Default is 1.}
}
\value{
Returns a list of 2 components : \describe{
\item{hat_beta_l}{lower bound estimates of regression coefficients}

\item{hat_beta_u}{upper bound estimates of regression coefficients}}
}
\description{
This function runs a two sample least squares when auxiliary data contains every right-hand side regressor
and main data contains a dependent variable and every right-hand side regressor but one omitted variable.
}
\references{
Hwang, Yujung (2021). Bounding Omitted Variable Bias Using Auxiliary Data. Working Paper.
}
\author{
Yujung Hwang, \email{yujungghwang@gmail.com}
}
