% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bndovb_tuning.R
\name{bndovb_tuning}
\alias{bndovb_tuning}
\title{bndovb_tuning}
\usage{
bndovb_tuning(
  maindat,
  auxdat,
  depvar,
  ovar,
  comvar,
  method = 1,
  mainweights = NULL,
  auxweights = NULL,
  signres = NULL,
  nboot = 100,
  scalegrid = c(-1/2, -1/3, -1/4, -1/5, -1/6),
  tau = 0.05,
  seed = 210823,
  parallel = TRUE
)
}
\arguments{
\item{maindat}{Main data set. It must be a data frame.}

\item{auxdat}{Auxiliary data set. It must be a data frame.}

\item{depvar}{A name of a dependent variable in main dataset}

\item{ovar}{A name of an omitted variable in main dataset which exists in auxiliary data}

\item{comvar}{A vector of the names of common regressors existing in both main data and auxiliary data}

\item{method}{CDF and Quantile function estimation method.
Users can choose either 1 or 2. If the method is 1, the CDF and quantile function is estimated assuming a parametric normal distribution.
If the method is 2, the CDF and quantile function is estimated using a nonparaemtric estimator in Li and Racine(2008) \doi{10.1198/073500107000000250}, Li, Lin, and Racine(2013) \doi{10.1080/07350015.2012.738955}.
Default is 1.}

\item{mainweights}{An optional weight vector for the main dataset. The length must be equal to the number of rows of 'maindat'.}

\item{auxweights}{An optional weight vector for the auxiliary dataset. The length must be equal to the number of rows of 'auxdat'.}

\item{signres}{An option to impose a sign restriction on a coefficient of an omitted variable. Set either NULL or pos or neg.
Default is NULL. If NULL, there is no sign restriction.
If 'pos', the estimator imposes an extra restriction that the coefficient of an omitted variable must be positive.
If 'neg', the estimator imposes an extra restriction that the coefficient of an omitted variable must be negative.}

\item{nboot}{Number of bootstraps to compute the confidence interval. Default is 100.}

\item{scalegrid}{Tuning parameter grid to search. It must be a vector of numbers between -1/2 and 0. Default is c(-1/2,-1/3,-1/4,-1/5,-1/6).}

\item{tau}{Significance level. (1-tau)\% confidence interval is computed. Default is 0.05.}

\item{seed}{Seed for random number generation. Default is 210823.}

\item{parallel}{Either TRUE or FALSE. Whether to compute in parallel. Default is TRUE.}
}
\value{
Returns a list of 3 components : \describe{

\item{optimal_scale}{An optimal scale parameter which gives coverage rates closest to (1-tau)}

\item{cover_beta_l}{A matrix of coverage rates of the lower bound parameters under different scale parameters}

\item{cover_beta_u}{A matrix of coverage rates of the lower bound parameters under different scale parameters}}
}
\description{
This function computes an optimal tuning parameter to compute the confidence interval for bndovb function
The function returns an optimal tuning parameter using double bootstrap procedure
}
\examples{
data(maindat_nome)
data(auxdat_nome)

# To shorten computation time, I set the number of bootstrap small in an example below.
# In practice, please set it a large number
bndovb_tuning(maindat_nome,auxdat_nome,depvar="y",ovar="x1",comvar=c("x2","x3"),method=1,nboot=2)


}
\references{
\describe{
\item{Hwang, Yujung (2021)}{Bounding Omitted Variable Bias Using Auxiliary Data. Available at SSRN.\doi{10.2139/ssrn.3866876}}}
}
\author{
Yujung Hwang, \email{yujungghwang@gmail.com}
}
