% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bndovbme_tuning.R
\name{bndovbme_tuning}
\alias{bndovbme_tuning}
\title{bndovbme_tuning}
\usage{
bndovbme_tuning(
  maindat,
  auxdat,
  depvar,
  pvar,
  ptype = 1,
  comvar,
  sbar = 2,
  mainweights = NULL,
  auxweights = NULL,
  normalize = TRUE,
  signres = NULL,
  nboot = 100,
  scalegrid = c(-1/2, -1/3, -1/4, -1/5, -1/6),
  tau = 0.05,
  seed = 210823,
  parallel = TRUE
)
}
\arguments{
\item{maindat}{Main data set. It must be a data frame.}

\item{auxdat}{Auxiliary data set. It must be a data frame.}

\item{depvar}{A name of a dependent variable in main dataset}

\item{pvar}{A vector of the names of the proxy variables for the omitted variable.
When proxy variables are continuous, the first proxy variable is used as an anchoring variable.
When proxy variables are discrete, the first proxy variable is used for initialization (For details, see a documentation for "dproxyme" function).}

\item{ptype}{Either 1 (continuous) or 2 (discrete). Whether proxy variables are continuous or discrete. Default is 1 (continuous).}

\item{comvar}{A vector of the names of the common regressors existing in both main data and auxiliary data}

\item{sbar}{A cardinality of the support of the discrete proxy variables. Default is 2. If proxy variables are continuous, this variable is irrelevant.}

\item{mainweights}{An optional weight vector for the main dataset. The length must be equal to the number of rows of 'maindat'.}

\item{auxweights}{An optional weight vector for the auxiliary dataset. The length must be equal to the number of rows of 'auxdat'.}

\item{normalize}{Whether to normalize the omitted variable to have mean 0 and standard deviation 1. Set TRUE or FALSE.
Default is TRUE. If FALSE, then the scale of the omitted variable is anchored with the first proxy variable in pvar list.}

\item{signres}{An option to impose a sign restriction on a coefficient of an omitted variable. Set either NULL or pos or neg.
Default is NULL. If NULL, there is no sign restriction.
If 'pos', the estimator imposes an extra restriction that the coefficient of an omitted variable must be positive.
If 'neg', the estimator imposes an extra restriction that the coefficient of an omitted variable must be negative.}

\item{nboot}{Number of bootstraps to compute the confidence interval. Default is 100.}

\item{scalegrid}{Tuning parameter grid to search. It must be a vector of numbers between -1/2 and 0. Default is c(-1/2,-1/3,-1/4,-1/5,-1/6).}

\item{tau}{Significance level. (1-tau)\% confidence interval is computed. Default is 0.05.}

\item{seed}{Seed for random number generation. Default is 210823.}

\item{parallel}{Either TRUE or FALSE. Whether to compute in parallel. Default is TRUE.}
}
\value{
Returns a list of 3 components : \describe{

\item{optimal_scale}{An optimal scale parameter which gives coverage rates closest to (1-tau)}

\item{cover_beta_l}{A matrix of coverage rates of the lower bound parameters under different scale parameters}

\item{cover_beta_u}{A matrix of coverage rates of the lower bound parameters under different scale parameters}}
}
\description{
This function computes an optimal tuning parameter to compute the confidence interval for bndovbme function
The function returns an optimal tuning parameter using double bootstrap procedure
}
\examples{
## load example data
data(maindat_mecont)
data(auxdat_mecont)

## set ptype=1 for continuous proxy variables
pvar<-c("z1","z2","z3")
cvar<-c("x","w1")

# To shorten computation time, I set the number of bootstrap small in an example below.
# In practice, please set it a large number
bndovbme_tuning(maindat_mecont,auxdat_mecont,depvar="y",pvar=pvar,ptype=1,comvar=cvar,nboot=2)

}
\references{
\describe{
\item{Hwang, Yujung (2021)}{Bounding Omitted Variable Bias Using Auxiliary Data. Available at SSRN.\doi{10.2139/ssrn.3866876}}}
}
\author{
Yujung Hwang, \email{yujungghwang@gmail.com}
}
