% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizeNormDist.R
\name{discretizeNormDist}
\alias{discretizeNormDist}
\title{discretizeNormDist}
\usage{
discretizeNormDist(n, mean, var)
}
\arguments{
\item{n}{Number of grid points}

\item{mean}{Mean of a normal distribution to discretize}

\item{var}{Variance of a normal distribution to discretize}
}
\value{
Returns an equal probability grid point : \describe{
\item{EVbetweenZ}{n grid points with equal probability}}
}
\description{
This function discretizes a normal distribution into n grid points with equal probability
}
\examples{
discretizeNormDist(n=3,mean=0,var=1)

}
\references{
\describe{
\item{Adda, J., and Cooper, R. W. (2003)}{Dynamic economics: quantitative methods and applications. MIT press.}{Bounding Omitted Variable Bias Using Auxiliary Data. Working Paper.}}
}
\author{
Yujung Hwang, \email{yujungghwang@gmail.com}
}
