% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maindat_medisc}
\alias{maindat_medisc}
\title{A simulated main data to show how to use 'bndovbme' function with discrete proxy variables}
\format{
A data frame with 3000 rows and 3 variables:
\describe{
\item{w1}{A common covariate in both main and auxiliary data}
\item{x}{A common covariate in both main and auxiliary data}
\item{y}{A dependent variable}
}
}
\source{
This dataset was simulated by simulatePackageData.R in data-raw folder
}
\usage{
maindat_medisc
}
\description{
A simulated main data to show how to use 'bndovbme' function with discrete proxy variables
}
\keyword{datasets}
