% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maindat_nome}
\alias{maindat_nome}
\title{A simulated main data to show how to use 'bndovb' function}
\format{
A data frame with 100000 rows and 3 variables:
\describe{
\item{x2}{A common covariate in both main and auxiliary data}
\item{x3}{A common covariate in both main and auxiliary data}
\item{y}{A dependent variable}
}
}
\source{
This dataset was simulated by simulatePackageData.R in data-raw folder
}
\usage{
maindat_nome
}
\description{
A simulated main data to show how to use 'bndovb' function
}
\keyword{datasets}
