% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bndovbme.R
\name{bndovbme}
\alias{bndovbme}
\title{bndovbme}
\usage{
bndovbme(
  maindat,
  auxdat,
  depvar,
  pvar,
  ptype = 1,
  comvar,
  sbar = 2,
  coefub = 100,
  coeflb = -100
)
}
\arguments{
\item{maindat}{Main data set}

\item{auxdat}{Auxiliary data set}

\item{depvar}{A name of a dependent variable in main dataset}

\item{pvar}{A vector of the names of the proxy variables for the omitted variable.
When proxy variables are continuous, the first proxy variable is used as an anchoring variable.
When proxy variables are discrete, the first proxy variable is used for initialization (For details, see a documentation for "dproxyme" function).}

\item{ptype}{Either 1 (continuous) or 2 (discrete). Whether proxy variables are continuous or discrete. Default is 1 (continuous).}

\item{comvar}{A vector of the names of the common regressors existing in both main data and auxiliary data}

\item{sbar}{A cardinality of the support of the discrete proxy variables. Default is 2. If proxy variables are continuous, this variable is irrelevant.}

\item{coefub}{An upper bound constraint in the Maximum Likelihood estimation. Default is 100. If you do not want to set any constraint, just set it to a very large number.}

\item{coeflb}{A lower bound constraint in the Maximum Likelihood estimation coefficients. Default is -100. If you do not want to set any constraint, just set it to a very small number.}
}
\value{
Returns a list of 2 components : \describe{
\item{hat_beta_l}{lower bound estimates of regression coefficients}

\item{hat_beta_u}{upper bound estimates of regression coefficients}}
}
\description{
This function runs a two sample least squares when main data contains a dependent variable and
every right hand side regressor but one omitted variable.
The function requires an auxiliary data which includes every right hand side regressor but one omitted variable,
and enough proxy variables for the omitted variable.
When the omitted variable is continuous, the auxiliary data must contain at least two continuous proxy variables.
When the omitted variable is discrete, the auxiliary data must contain at least three continuous proxy variables.
}
\references{
Hwang, Yujung (2021). Bounding Omitted Variable Bias Using Auxiliary Data. Working Paper.
}
\author{
Yujung Hwang, \email{yujungghwang@gmail.com}
}
